requirements.txt
streamlit
pyyaml
requests
gitpython
streamlit-file-selector
